feature/ modifying id type and dockerfile, notification controller

+++ src/main/java/com/example/api_spring/models/Anuncio.java
@@ -17,7 +17,7 @@ public class Anuncio {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_anuncio")
-    private int idAnuncio;
+    private Long idAnuncio;
 
     private double preco;
