+++ src/main/java/com/example/api_spring/repositories/PostagemRepository.java
@@ -3,5 +3,5 @@ package com.example.api_spring.repositories;
 import com.example.api_spring.models.Postagem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostagemRepository extends JpaRepository<Postagem,Integer> {
+public interface PostagemRepository extends JpaRepository<Postagem,Long> {
 }
